pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepseek4pdf"
version = "0.1.0"
description = "A simple library for querying PDFs using LLMs with DeepSeek R1"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Vibhatha Abeykoon", email = "vibhatha@gmail.com" },
]
dependencies = [
    "llama-index-core",
    "llama-index-llms-ollama",
    "llama-index-embeddings-huggingface",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/vibhatha/deepseek4pdf"
Documentation = "https://github.com/vibhatha/deepseek4pdf#readme"
Repository = "https://github.com/vibhatha/deepseek4pdf.git"
Issues = "https://github.com/vibhatha/deepseek4pdf/issues"

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.3.0",
]

[tool.black]
line-length = 88
target-version = ["py37"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=deepseek4pdf"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["deepseek4pdf"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]